organized test folder and updated imports
