--- ResumeUpload/middleware/__mocks__/authentication.js
@@ -1,3 +1,0 @@
-module.exports = {
-  authenticateApiKey: jest.fn(() => true),
-};
