+++ shared/middleware/__mocks__/authentication.js
