+++ shared/middleware/authentication.js
