+++ tests/RequestTrigger/__mocks__/applicationinsights.js
